@@ -3,38 +3,34 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.13.0",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.23",
-    "@types/js-cookie": "^3.0.1",
-    "@types/node": "^12.20.15",
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.7",
-    "axios": "^0.24.0",
-    "http-proxy-middleware": "^0.19.1",
-    "js-cookie": "^3.0.1",
-    "jwt-decode": "^3.1.2",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.14",
+    "@types/js-cookie": "^3.0.6",
+    "@types/node": "^22.9.0",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@vidstack/react": "1.12.12",
+    "axios": "^1.7.7",
+    "js-cookie": "^3.0.5",
+    "jwt-decode": "^4.0.0",
     "normalize.css": "^8.0.1",
-    "plyr": "3.6.8",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.6.1",
-    "react-cookie": "^4.0.3",
-    "react-device-detect": "^1.17.0",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-progressive-graceful-image": "^0.6.14",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "5.0.1",
-    "typescript": "^4.3.2",
-    "typescript-json-serializer": "^3.4.0",
-    "use-resize-observer": "^8.0.0"
+    "react": "^18.3.1",
+    "react-cookie": "^7.2.2",
+    "react-device-detect": "^2.2.3",
+    "react-dom": "^18.3.1",
+    "react-error-boundary": "^4.1.2",
+    "react-icons": "^5.3.0",
+    "react-progressive-graceful-image": "^0.7.0",
+    "react-router-dom": "^6.28.0",
+    "typescript": "^4.9.5",
+    "typescript-json-serializer": "^6.0.1",
+    "use-resize-observer": "^9.1.0"
   },
   "scripts": {
-    "start": "BROWSER=none react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview"
   },
   "eslintConfig": {
     "extends": [
@@ -55,10 +51,12 @@
     ]
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.176",
-    "@types/react-router-dom": "^5.1.7",
-    "gulp-autoprefixer": "^8.0.0",
-    "node-sass": "^7.0.0",
-    "sass-loader": "^12.4.0"
+    "@types/lodash": "^4.17.13",
+    "@types/react-router-dom": "^5.3.3",
+    "@vitejs/plugin-react": "^4.3.3",
+    "sass": "^1.81.0",
+    "sass-loader": "^16.0.3",
+    "vite": "^5.4.11",
+    "webpack": "^5.96.1"
   }
 }
