@@ -1,23 +1,26 @@
-import { BrowserRouter, Route, Switch, useParams } from 'react-router-dom';
-import Editor from './pages/Editor';
-import Compilation from './pages/Compilation';
-import Main from './pages/Main';
-import VideoWall from './pages/VideoWall';
+import {BrowserRouter, Route, Routes, useParams} from 'react-router-dom';
+import React, { Suspense, lazy } from 'react';
+
+const Editor = lazy(() => import('./pages/Editor'));
+const Compilation = lazy(() => import('./pages/Compilation'));
+const Main = lazy(() => import('./pages/Main'));
+const VideoWall = lazy(() => import('./pages/VideoWall'));
 
 function App() {
 
   return (
     <div className="app-root">
       <BrowserRouter>
-          <Switch>
-            <Route exact path="/" component = { Main }  />
-            <Route path="/search"  component = { Main } />
-            <Route path="/editor/:id" children = { <EditorRouter /> } />
-            <Route exact path="/video-wall" component = { VideoWall } />
-            <Route exact path="/compilation" children ={ <Compilation /> } />
-          </Switch>
+        <Suspense fallback = { <div>Loading...</div> }>
+          <Routes>
+            <Route path="/" element={<Main />} />
+            <Route path="/search" element={<Main />} />
+            <Route path="/editor/:id" element={<EditorRouter />} />
+            <Route path="/video-wall" element={<VideoWall />} />
+            <Route path="/compilation" element={<Compilation />} />
+          </Routes>
+        </Suspense>
       </BrowserRouter>
-
     </div>
   );
 }
@@ -25,7 +28,9 @@ function App() {
 function EditorRouter() {
   let { id } = useParams<{ id: string }>();
   return (
-    <Editor videoId={id} />
+      <>
+          { id && <Editor videoId = { id } /> }
+      </>
   );
 }
 
