@@ -1,17 +1,17 @@
-import { CSSProperties, useEffect, useRef, useState } from "react"
-import { FaSort } from "react-icons/fa"
-import { FiEdit, FiPlusCircle } from "react-icons/fi"
-import { IoCutSharp } from "react-icons/io5"
-import { AiOutlineDelete } from "react-icons/ai"
+import {CSSProperties, useEffect, useRef, useState} from "react"
+import {FaSort} from "react-icons/fa"
+import {FiEdit} from "react-icons/fi"
 import ProgressiveImage from "react-progressive-graceful-image"
-import { Api } from "../api/Api"
-import { ResourceSelection, SearchResult, Resource, ResourceUserMeta } from "../api/Model"
-import { dateMillisToString, formatByteSize } from "../api/Util"
+import {Api} from "../api/Api"
+import {Resource, ResourceSelection, ResourceUserMeta, SearchResult} from "../api/Model"
+import {dateMillisToString, formatByteSize} from "../api/Util"
 import './ListView.scss'
 import Scrollable from "./common/Scrollable"
-import { useSortParam } from "../api/Constants"
-import { useHistory } from "react-router-dom"
+import {useSortParam} from "../api/Constants"
+import {useNavigate} from "react-router-dom"
 import TagsBar from "./common/TagsBar"
+import {MdDelete, MdMovieEdit} from "react-icons/md";
+import {FaHashtag} from "react-icons/fa6";
 
 type ListProps = {
   selection: ResourceSelection
@@ -26,7 +26,8 @@ const ListView = (props: ListProps) => {
   const [searchResult, setSearchResult] = useState(initialSearchResult);
   const [isFetching, setIsFetching] = useState(false)
   const [fetchMore, setFetchMore] = useState(true)
-  const history = useHistory();
+  const [selectedItems, setSelectedItems] = useState<Array<number>>([])
+  const navigate = useNavigate();
 
   const fetchData = (previous: Array<Resource>) => {
 
@@ -58,84 +59,130 @@ const ListView = (props: ListProps) => {
 
   useEffect(() => { if (isFetching && fetchMore) fetchData(searchResult.results); }, [isFetching]);
 
+  const toggle = (index: number)=>  {
+    const i = selectedItems.indexOf(index)
+    if (i > -1) {
+      setSelectedItems(selectedItems.filter((v) => v !== index))
+    } else {
+      setSelectedItems([...selectedItems, index])
+    }
+  }
+
+  const headers =
+    <tr key="row-header" className="list-row">
+      <th className="list-header-select"><input type="checkbox"/></th>
+      <th className="list-header-thumbnail"></th>
+      <th className="list-header-title"><span>Title</span>
+        <FaSort className="column-sort-icon"
+                onClick={() => setSort({field: "title", direction: sort.direction === "asc" ? "desc" : "asc"})}/>
+      </th>
+      <th className="list-header-tags"><span>Tags</span></th>
+      <th className="list-header-date"><span>Date</span>
+        <FaSort className="column-sort-icon" onClick={() => setSort({
+          field: "date_added",
+          direction: sort.direction === "asc" ? "desc" : "asc"
+        })}/>
+      </th>
+      <th className="list-header-size"><span>Size</span>
+        <FaSort className="column-sort-icon"
+                onClick={() => setSort({field: "size", direction: sort.direction === "asc" ? "desc" : "asc"})}/>
+      </th>
+      <th className="list-header-resolution"><span>Quality</span>
+        {/* <FaSort className="column-sort-icon" onClick = { () => setSort({field: "resolution", direction: sort.direction === "asc" ? "desc" : "asc" }) } /> */}
+        {/* <BsThreeDotsVertical className="list-menu-icon" /> */}
+      </th>
+    </tr>
+
+  const actionBar =
+    <tr key="row-header" className="list-row">
+      <th className = "list-header-select"><input type="checkbox"/></th>
+      <th className = "list-header-actionbar" colSpan={6}>
+        <FaHashtag className= "action-bar-item" />
+        <MdDelete className = "action-bar-item" />
+      </th>
+    </tr>
+
   return (
-      <Scrollable
-        className = "list-container"
-        fetchContent = { () => { if (!isFetching && fetchMore) setIsFetching(true); fetchData(searchResult.results) } }
-        scrollType = 'page'
-      >
-      <div key="row-header" className="list-row">
-        {/* <div className="list-cell list-header list-select"><input type="checkbox" /></div> */}
-        <div className="list-cell list-header"></div>
-        <div className="list-cell list-header">Title
-          <FaSort className="column-sort-icon" onClick = { () => setSort({field: "title", direction: sort.direction === "asc" ? "desc" : "asc" }) } />
-        </div>
-        <div className="list-cell list-header">Tags</div>
-        <div className="list-cell list-header">Date
-          <FaSort className="column-sort-icon" onClick = { () => setSort({field: "date_added", direction: sort.direction === "asc" ? "desc" : "asc" }) } />
-        </div>
-        <div className="list-cell list-header">Size
-          <FaSort className="column-sort-icon" onClick = { () => setSort({field: "size", direction: sort.direction === "asc" ? "desc" : "asc" }) } />
-        </div>
-        <div className="list-cell list-header">
-          {/* <FaSort className="column-sort-icon" onClick = { () => setSort({field: "resolution", direction: sort.direction === "asc" ? "desc" : "asc" }) } /> */}
-          {/* <BsThreeDotsVertical className="list-menu-icon" /> */}
-        </div>
-      </div>
-      
-      <div key="row-spacer" className="list-row row-spacer"></div>
-      {
-        searchResult.results.map((v, index) => {
-          return(
-            <div key={`row-${v.id}`} className="list-row">
+    <Scrollable
+      className="list-container"
+      fetchContent={() => {
+        if (!isFetching && fetchMore) setIsFetching(true);
+        fetchData(searchResult.results)
+      }}
+      scrollType='page'
+    >
+      <tr key="row-column-width-spacer" style ={ {height: 0 } }>
+        <td style={{width: 36}}></td>
+        <td style={{width: 72}}></td>
+        <td style={{width: "65%"}}></td>
+        <td style={{width: "35%"}}></td>
+        <td style={{width: 110}}></td>
+        <td style={{width: 100}}></td>
+        <td style={{width: 80}}></td>
+      </tr>
 
-              <div key="thumbnail" className="list-cell list-thumbnail">
-                <ProgressiveImage src={v.urls.thumbnailUrl} placeholder="/image_placeholder.svg">
-                    { (src: string) => 
-                      <img className="list-thumbnail-img" src={src} onClick={() => props.onClick(v) } alt="an image" /> }
-                </ProgressiveImage>
-              </div>
+      {selectedItems.length > 0 ? actionBar : headers}
 
-              <TitleCell resource= { v } />
-              
-              <TagsCell resource= { v } />
+      <tr key="row-spacer" style = { { height : 4 } } />
+        {
+          searchResult.results.map((resource, index) => {
+            return (
+              <tr key={`row-${resource.resourceId}`} className="list-row">
 
-              <div key="date" className="list-cell list-date">
-                { dateMillisToString(v.uploadTimestamp) }
-              </div>
+                <td key="select" className="list-select" onClick={() => { toggle(index) }}>
+                  <input type="checkbox" checked={selectedItems.indexOf(index) > -1}/>
+                </td>
 
-              <div key="size" className="list-cell list-size">
-                  { formatByteSize(v.resourceInfo.sizeInBytes, 1) }
-              </div>
+                <td key="thumbnail" className="list-thumbnail">
+                  <ProgressiveImage src={resource.urls.thumbnailUrl} placeholder="/image_placeholder.svg">
+                    {(src: string) =>
+                      <img className="list-thumbnail-img" src={src} onClick={() => props.onClick(resource)}
+                           alt="an image"/>}
+                  </ProgressiveImage>
+                </td>
 
-              <div key="resolution" className="list-cell list-resolution">
-                <div className = "cell-wrapper">
-                { 
-                  Api.session().isAdmin() && 
-                    <div className = "media-actions">
-                      <IoCutSharp className = "fragments-action" onClick = { () => history.push(`/editor/${v.id}`) } />
-                      <AiOutlineDelete className = "delete-action" />
-                    </div> 
-                }
-                { `${v.resourceMeta.height}p` }
-                </div>
-                
-              </div>
-            </div>  
-          );
-        })
-      }
-      </Scrollable>
-  );
+                <TitleCell mediaResource={resource} onClick={() => { toggle(index) }}/>
+
+                <TagsCell resource={resource}/>
+
+                <td key="date" className="list-cell list-date">
+                  {dateMillisToString(resource.uploadTimestamp)}
+                </td>
+
+                <td key="size" className="list-cell list-size">
+                  {formatByteSize(resource.resourceInfo.sizeInBytes, 1)}
+                </td>
+
+                <td key="resolution" className="list-cell list-resolution">
+                  <div className="cell-wrapper">
+                    {
+                      Api.session().isAdmin() &&
+                        <div className="media-actions">
+                            <MdMovieEdit className="fragments-action"
+                                         onClick={() => navigate(`/editor/${resource.resourceId}`)}/>
+                            <MdDelete className="delete-action"/>
+                        </div>
+                    }
+                    {`${resource.resourceMeta.height}p`}
+                  </div>
+
+                </td>
+              </tr>
+            );
+          })
+        }
+    </Scrollable>
+);
 }
 
-const TagsCell = (props: {resource: Resource }) => {
+const TagsCell = (props: {
+  resource: Resource }) => {
   const [tags, setTags] = useState(props.resource.userMeta.tags)
   const isAdmin = Api.session().isAdmin()
 
   const updateTags = (newTags: Array<string>) => {
     const meta: ResourceUserMeta = { ...props.resource.userMeta, tags: newTags }
-    Api.updateUserMetaData(props.resource.bucketId, props.resource.id, meta).then(() =>  {
+    Api.updateUserMetaData(props.resource.bucketId, props.resource.resourceId, meta).then(() =>  {
       setTags(newTags)
     })
   }
@@ -146,9 +193,11 @@ const TagsCell = (props: {resource: Resource }) => {
             showDeleteButton = {isAdmin} />
 }
 
-const TitleCell = (props: { resource: Resource} ) => {
+type TitleCellProps =  { mediaResource: Resource; } & React.HTMLProps<HTMLTableCellElement>;
 
-  const [title, setTitle] = useState(props.resource.userMeta.title)
+const TitleCell = ({ mediaResource, ...elementProps }: TitleCellProps ) => {
+
+  const [title, setTitle] = useState(mediaResource.userMeta.title)
   const [editTitle, setEditTitle] = useState(false)
   const inputRef = useRef<HTMLInputElement>(null)
 
@@ -158,8 +207,8 @@ const TitleCell = (props: { resource: Resource} ) => {
   }, [editTitle])
 
   const updateTitle = (newTitle: string) => {
-    const meta: ResourceUserMeta = { ...props.resource.userMeta, title: newTitle }
-    Api.updateUserMetaData(props.resource.bucketId, props.resource.id, meta).then(() =>  {
+    const meta: ResourceUserMeta = { ...mediaResource.userMeta, title: newTitle }
+    Api.updateUserMetaData(mediaResource.bucketId, mediaResource.resourceId, meta).then(() =>  {
       setTitle(newTitle)
       setEditTitle(false)
     })
@@ -168,17 +217,18 @@ const TitleCell = (props: { resource: Resource} ) => {
   const style: CSSProperties = editTitle ? { paddingLeft: "3px"} : {}
 
   return(
-    <div style = { style } key="title" className="list-cell list-title">
+    <td style = { style } key="title" className="list-cell list-title" {...elementProps }>
       <div className="cell-wrapper">
         { !editTitle && title }
-        { (!editTitle && Api.session().isAdmin()) && <FiEdit onClick = { () => { setEditTitle(true); } } className="edit-title action-icon hover-action" /> }
+        { (!editTitle && Api.session().isAdmin()) &&
+            <FiEdit onClick = { () => { setEditTitle(true); } } className="edit-title" /> }
         { editTitle && 
           <input 
             ref        = { inputRef } 
             className  = "edit-title-input"
             type       = "text" 
             value      = { title } 
-            onBlur     = { () => { setEditTitle(false) } } 
+            onBlur     = { () => { setEditTitle(false) } }
             onChange   = { (e) => { setTitle(e.target.value ) } }
             onKeyPress = { (e) => {
               if (e.key === "Enter") {
@@ -189,7 +239,7 @@ const TitleCell = (props: { resource: Resource} ) => {
           />
         }
       </div>
-    </div>
+    </td>
   );
 }
 
