@@ -8,11 +8,13 @@ class AuthServiceImpl(config: AuthConfig) extends AuthServiceGrpc.AuthService {
 
   private val tokenManager = new TokenManager(config.jwt)
   private val adminUserId = "0"
+  
+  // https://github.com/Password4j/password4j
 
   override def login(request: Credentials): Future[LoginResponse] = {
 
     if (request.username == config.adminUsername && request.password == config.adminPassword)
-      Future.successful(api.Authentication(adminUserId, tokenManager.createToken(adminUserId)))
+      Future.successful(api.Authentication(adminUserId, tokenManager.createJwtToken(adminUserId, Set("admin"))))
     else
       Future.successful(api.InvalidCredentials())
   }
