@@ -6,14 +6,15 @@ import java.time.Instant
 
 class TokenManager(jwtConfig: JwtConfig) {
 
-  private val expirationInSeconds = jwtConfig.tokenExpiration.toSeconds
+  private val expirationInSeconds = jwtConfig.accessTokenExpiration.toSeconds
 
-  def createToken(userId: String): String = {
+  def createJwtToken(userId: String, roles: Set[String]): String = {
 
     val claim = JwtClaim(
       expiration = Some(Instant.now.plusSeconds(expirationInSeconds).getEpochSecond),
-      issuedAt   = Some(Instant.now.getEpochSecond)
-    ) + ("userId", userId) + ("admin", true)
+      issuedAt   = Some(Instant.now.getEpochSecond),
+      notBefore  = Some(Instant.now.getEpochSecond),
+    ) + ("userId", userId) + ("roles", roles)
 
     val token = JwtCirce.encode(claim, jwtConfig.secretKey, jwtConfig.algo)
 
