@@ -1,9 +1,10 @@
 package nl.amony.service.resources
 
-import nl.amony.lib.hash.Base32
-import nl.amony.lib.hash.PartialHash.partialHash
-import pureconfig._
-import pureconfig.generic.derivation.default._
+import cats.effect.IO
+import nl.amony.service.resources.util.Base32
+import nl.amony.service.resources.util.PartialHash.partialHash
+import pureconfig.*
+import pureconfig.generic.derivation.default.*
 
 import java.nio.file.Path
 import java.security.MessageDigest
@@ -12,22 +13,28 @@ import scala.concurrent.duration.FiniteDuration
 object ResourceConfig {
 
   sealed trait ResourceBucketConfig derives ConfigReader
+  
+  case class ScanConfig(
+     enabled: Boolean,
+     scanParallelFactor: Int,
+     pollInterval: FiniteDuration,
+     verifyExistingHashes: Boolean,
+     hashingAlgorithm: HashingAlgorithm,
+     extensions: List[String],
+  ) derives ConfigReader
 
   case class LocalDirectoryConfig(
-      id: String,
-      private val path: Path,
-      scanParallelFactor: Int,
-      verifyExistingHashes: Boolean,
-      hashingAlgorithm: HashingAlgorithm,
-      relativeResourcePath: Path,
-      extensions: List[String]
+     id: String,
+     private val path: Path,
+     scan: ScanConfig,
+     relativeCachePath: Path,
+     relativeUploadPath: Path,
+     
   ) extends ResourceBucketConfig {
 
-    lazy val writePath: Path    = path.toAbsolutePath.normalize().resolve(relativeResourcePath)
+    lazy val cachePath: Path    = path.toAbsolutePath.normalize().resolve(relativeCachePath)
     lazy val resourcePath: Path = path.toAbsolutePath.normalize()
-
-    def filterFileName(fileName: String): Boolean =
-      extensions.exists(ext => fileName.endsWith(s".$ext")) && !fileName.startsWith(".")
+    lazy val uploadPath: Path   = path.toAbsolutePath.normalize().resolve(relativeUploadPath)
   }
 
   case class TranscodeSettings(
@@ -38,21 +45,22 @@ object ResourceConfig {
 
   sealed trait HashingAlgorithm derives ConfigReader {
     def algorithm: String
-    def createHash(path: Path): String
+    def createHash(path: Path): IO[String]
     def encodeHash(bytes: Array[Byte]): String
   }
 
   case object PartialHash extends HashingAlgorithm {
     override val algorithm = "SHA-1"
-    override def createHash(path: Path): String =
+    override def createHash(path: Path): IO[String] =
       partialHash(
-        file    = path,
-        nBytes  = 512,
-        hasher  = () => { MessageDigest.getInstance(algorithm) },
-        encoder = encodeHash
+        file      = path,
+        nChunks   = 32,
+        chunkSize = 32,
+        hasher    = () => MessageDigest.getInstance(algorithm),
+        encoder   = encodeHash
       )
 
     override def encodeHash(bytes: Array[Byte]): String =
-      Base32.encode(bytes).substring(0, 24)
+      Base32.encode(bytes).substring(0, 24) // this is a 24 character hash base 32 = 120 bits
   }
 }
