@@ -1,45 +1,36 @@
 package nl.amony.service.resources.web
 
-import io.circe.generic.semiauto.deriveCodec
-import io.circe.{Codec, Decoder, Encoder}
-import nl.amony.service.resources.api.{ImageMeta, VideoMeta, ResourceMeta}
+import io.circe.Encoder
+import nl.amony.service.resources.api.{ImageMeta, ResourceInfo, ResourceMeta, VideoMeta}
 import nl.amony.service.resources.web.ResourceWebModel.*
 
 object JsonCodecs {
-
-  // web model codecs
-  given fragmentCodec: Codec[FragmentDto]      = deriveCodec[FragmentDto]
-  given mediaInfoCodec: Codec[ResourceMetaDto] = deriveCodec[ResourceMetaDto]
-  given urlsCodec: Codec[ResourceUrlsDto]      = deriveCodec[ResourceUrlsDto]
-  given codec: Codec[ResourceInfoDto]          = deriveCodec[ResourceInfoDto]
-  given videoMetaCodec: Codec[UserMetaDto]     = deriveCodec[UserMetaDto]
-  given resourceEncoder: Codec[ResourceDto]    = deriveCodec[ResourceDto]
-
+  
   // contra map encoders for internal protocol classes
-  given resourceInfoEncoder: Encoder[nl.amony.service.resources.api.ResourceInfo] =
-    resourceEncoder.contramap[nl.amony.service.resources.api.ResourceInfo](toDto)
+  given resourceInfoEncoder: Encoder[ResourceInfo] =
+    summon[Encoder[ResourceDto]].contramap[ResourceInfo](toDto)
 
-  def toDto(resource: nl.amony.service.resources.api.ResourceInfo): ResourceDto = {
+  def toDto(resource: ResourceInfo): ResourceDto = {
 
-    val resolutions: List[Int] = (resource.height :: List(320)).sorted
+    val resolutions: List[Int] = (resource.height :: List(352)).sorted
+
+    val thumbnailTimestamp: Long = resource.thumbnailTimestamp.getOrElse(resource.durationInMillis() / 3)
 
     val urls = {
-
-      val thumbnailTimestamp = resource.durationInMillis() / 3
 
       val tsPart = if (thumbnailTimestamp != 0) s"_${thumbnailTimestamp}" else ""
 
       ResourceUrlsDto(
-        originalResourceUrl  = s"/resources/${resource.bucketId}/${resource.hash}.mp4",
-        thumbnailUrl         = s"/resources/${resource.bucketId}/${resource.hash}${tsPart}_${resolutions.min}p.webp",
-        previewThumbnailsUrl = Some(s"/resources/${resource.bucketId}/${resource.hash}-timeline.vtt")
+        originalResourceUrl  = s"/api/resources/${resource.bucketId}/${resource.hash}/content",
+        thumbnailUrl         = s"/api/resources/${resource.bucketId}/${resource.hash}/thumb${tsPart}_${resolutions.min}p.webp",
+        previewThumbnailsUrl = Some(s"/api/resources/${resource.bucketId}/${resource.hash}/timeline.vtt")
       )
     }
 
     val meta = UserMetaDto(
-      title   = resource.title.orElse(Some(resource.fileName())),
+      title       = resource.title.orElse(Some(resource.fileName())),
       description = resource.description,
-      tags    = resource.tags.toList
+      tags        = resource.tags.toList
     )
 
     val mediaInfo: ResourceMetaDto = resource.contentMeta match {
@@ -49,14 +40,16 @@ object JsonCodecs {
             height    = height,
             duration  = 0,
             fps       = 0,
+            codec     = None,
           )
 
-      case VideoMeta(width, height, fps, duration, _) =>
+      case VideoMeta(width, height, fps, duration, codec, _) =>
           ResourceMetaDto(
             width     = width,
             height    = height,
             duration  = duration,
             fps       = fps,
+            codec     = codec,
           )
 
       case ResourceMeta.Empty =>
@@ -65,22 +58,23 @@ object JsonCodecs {
             height    = 0,
             duration  = 0,
             fps       = 0,
+            codec     = None,
           )
     }
 
     val resourceInfo = ResourceInfoDto(
       sizeInBytes = resource.size,
-      hash = resource.hash
+      hash        = resource.hash,
+      path        = resource.path
     )
 
     // hard coded for now
-    val start = (mediaInfo.duration / 3)
-    val range = (start, Math.min(mediaInfo.duration, start + 3000))
-    val highlights = List(FragmentDto(resource.hash, 0, range, List.empty, None, List.empty))
+    val range = (thumbnailTimestamp, Math.min(mediaInfo.duration, thumbnailTimestamp + 3000))
+    val highlights = List(ClipDto(resource.hash, range, List.empty, None, List.empty))
 
     ResourceDto(
       bucketId  = resource.bucketId,
-      id        = resource.hash,
+      resourceId = resource.hash,
       uploader  = "0",
       uploadTimestamp = resource.getCreationTime,
       urls = urls,
@@ -88,22 +82,21 @@ object JsonCodecs {
       contentType = resource.contentType.getOrElse("unknown"),
       resourceMeta = mediaInfo,
       resourceInfo = resourceInfo,
-      highlights = {
-        highlights.zipWithIndex.map { case (f, index) =>
+      thumbnailTimestamp = resource.thumbnailTimestamp,
+      clips = {
+        highlights.map { f =>
 
           val (start, end) = f.range
-
           val urls = resolutions.map(height =>
-            s"/resources/${resource.bucketId}/${resource.hash}~${start}-${end}_${height}p.mp4"
+            s"/api/resources/${resource.bucketId}/${resource.hash}/clip_${start}-${end}_${height}p.mp4"
           )
 
-          FragmentDto(
-            media_id = resource.hash,
-            index    = index,
-            range    = (start, end),
-            urls     = urls,
-            comment  = f.comment,
-            tags     = f.tags
+          ClipDto(
+            resourceId   = resource.hash,
+            range        = (start, end),
+            urls         = urls,
+            description  = f.description,
+            tags         = f.tags
           )
         }
       }
