@@ -1,34 +1,42 @@
 package nl.amony.service.resources.web
 
+import io.circe.*
+
 object ResourceWebModel {
+  
+  case class ThumbnailTimestampDto(
+     timestampInMillis: Long
+  ) derives Encoder.AsObject, Decoder
 
   case class UserMetaDto(
     title: Option[String],
     description: Option[String],
     tags: List[String]
-  )
+  ) derives Encoder.AsObject, Decoder
 
   case class ResourceMetaDto(
     width: Int,
     height: Int,
-    fps: Double,
+    fps: Float,
     duration: Long,
-  )
+    codec: Option[String],  
+  ) derives Encoder.AsObject
 
   case class ResourceUrlsDto(
     originalResourceUrl: String,
     thumbnailUrl: String,
     previewThumbnailsUrl: Option[String],
-  )
+  ) derives Encoder.AsObject
 
   case class ResourceInfoDto(
     hash: String,
-    sizeInBytes: Long
-  )
+    sizeInBytes: Long,
+    path: String
+  ) derives Encoder.AsObject
 
   case class ResourceDto(
     bucketId: String,
-    id: String,
+    resourceId: String,
     uploader: String,
     uploadTimestamp: Long,
     userMeta: UserMetaDto,
@@ -36,15 +44,15 @@ object ResourceWebModel {
     resourceMeta: ResourceMetaDto,
     resourceInfo: ResourceInfoDto,
     urls: ResourceUrlsDto,
-    highlights: List[FragmentDto],
-  )
+    thumbnailTimestamp: Option[Long],
+    clips: List[ClipDto],
+  ) derives Encoder.AsObject
 
-  case class FragmentDto(
-     media_id: String,
-     index: Int,
-     range: (Long, Long),
-     urls: List[String],
-     comment: Option[String],
-     tags: List[String]
-   )
+  case class ClipDto(
+    resourceId: String,
+    range: (Long, Long),
+    urls: List[String],
+    description: Option[String],
+    tags: List[String]
+   ) derives Encoder.AsObject
 }
