@@ -18,21 +18,20 @@ export type ResourceInfo = {
 }
 
 export type Resource = {
-  id: string,
+  resourceId: string,
   bucketId: String,
   uploader: string,
   uploadTimestamp: number,
-  // the media info of the originally uploaded file
   userMeta: ResourceUserMeta,
   contentType: String,
   resourceMeta: ResourceMeta,
   resourceInfo: ResourceInfo,
   urls: ResourceUrls,
-  highlights: Array<Fragment>,
+  clips: Array<Clip>,
 }
 
-export type Fragment = {
-  media_id: string,
+export type Clip = {
+  resourceId: string,
   range: [number, number],
   index: number,
   urls: string[],
@@ -85,5 +84,6 @@ export type Prefs = {
   showTitles: boolean
   showDuration: boolean
   showDates: boolean
+  showResolution: boolean
   gallery_columns: Columns
 }
