@@ -1,15 +1,11 @@
-import React, { CSSProperties, useEffect, useState } from 'react';
+import React, {CSSProperties, useEffect, useState} from 'react';
 import ProgressiveImage from "react-progressive-graceful-image";
-import { Api } from "../api/Api";
-import { Resource } from "../api/Model";
+import {Resource} from "../api/Model";
 import {dateMillisToString, durationInMillisToString, labelForResolution} from "../api/Util";
-import Dialog from './common/Dialog';
-import { DropDown, MenuItem } from './common/DropDown';
 import FragmentsPlayer from "./common/FragmentsPlayer";
 import ImgWithAlt from "./common/ImgWithAlt";
-import Modal from './common/Modal';
-import MediaInfo from './dialogs/MediaInfo';
 import './Preview.scss';
+import {ErrorBoundary} from "react-error-boundary";
 
 export type PreviewProps = {
   resource: Resource,
@@ -26,41 +22,34 @@ export type PreviewOptions = {
   showInfoBar: boolean,
   showDates: boolean,
   showDuration: boolean,
+  showResolution: boolean,
   showMenu: boolean,
 }
 
 const Preview = (props: PreviewProps) => {
   const [resource, setResource] = useState(props.resource)
   const [isHovering, setIsHovering] = useState(false)
-  const [showVideoPreview, setShowVideoPreview] = useState(false)
+  // const [showVideoPreview, setShowVideoPreview] = useState(false)
 
   const durationStr = durationInMillisToString(resource.resourceMeta.duration)
 
   const isVideo = resource.contentType.startsWith("video")
 
-  useEffect(() => {
-    setShowVideoPreview(isHovering)
-  }, [isHovering])
+  // useEffect(() => {
+  //   setShowVideoPreview(isHovering)
+  // }, [isHovering])
 
   const titlePanel =
       <div className = "preview-info-bar">
         <span className="media-title" title={resource.userMeta.title}>{resource.userMeta.title}</span>
         { props.options.showDates && <span className="media-date">{dateMillisToString(resource.uploadTimestamp)}</span> }
-        { !props.options.showDates && <span className="media-date">{`${resource.resourceMeta.height}p` }</span>}
       </div>
 
   const overlay =
       <div className="preview-overlay">
-        {
-            (props.options.showMenu && isHovering) &&
-            <div className = "preview-menu-container">
-              <PreviewMenu
-                  resource     = { resource }
-                  setVideo     = { setResource }
-                  onDialogOpen = { () => { setShowVideoPreview(false) } }/>
-            </div>
-        }
+        { props.options.showResolution && <div className="preview-quality-overlay">{labelForResolution(resource.resourceMeta.height)}</div> }
         { (isVideo && props.options.showDuration) && <div className="duration-overlay">{durationStr}</div> }
+        { isHovering && <a className="preview-edit-icon-overlay" href={`/editor/${props.resource.resourceId}`}><ImgWithAlt src="/icons/edit.svg" /></a> }
         {/* { <div className="abs-bottom-right"><FiDownload /></div> } */}
       </div>
 
@@ -77,19 +66,21 @@ const Preview = (props: PreviewProps) => {
 
   const videoPreview =
       <FragmentsPlayer
-          key       = { `video-preview-${props.resource.id}` }
+          key       = { `video-preview-${props.resource.resourceId}` }
           className = { `preview-video preview-media` }
           onClick   = { () => props.onClick(props.resource) }
-          fragments = { props.resource.highlights } />
+          fragments = { props.resource.clips } />
 
   const preview =
+    <ErrorBoundary fallback={ <div /> }>
       <div className    = "preview-media-container"
            onMouseEnter = { () => props.options.showPreviewOnHover && setIsHovering(true) }
            onMouseLeave = { () => setIsHovering(false) }>
-        { isVideo && showVideoPreview && videoPreview }
+        { isVideo && isHovering && videoPreview }
         { primaryThumbnail }
         { overlay }
       </div>
+    </ErrorBoundary>
 
   return (
       <div className = "preview-media">
@@ -97,65 +88,6 @@ const Preview = (props: PreviewProps) => {
         { props.options.showInfoBar && titlePanel }
       </div>
   )
-}
-
-const PreviewMenu = (props: {resource: Resource, setVideo: (v: Resource) => void, onDialogOpen: () => any}) => {
-
-  const [showDeleteDialog, setShowDeleteDialog] = useState(false);
-  const [showInfoModal, setShowInfoModal] = useState(false)
-
-  const cancelDelete = () => setShowDeleteDialog(false);
-  const confirmDelete = () => {
-    Api.deleteResourceById(props.resource.bucketId, props.resource.id).then(() => {
-      console.log("video was deleted")
-      setShowDeleteDialog(false)
-    })
-  };
-
-  return (
-    <>
-      <Modal visible = { showInfoModal } onHide = {() => setShowInfoModal(false)} >
-        <MediaInfo 
-          meta = { props.resource.userMeta }
-          onClose = { (meta) => {
-            Api.updateUserMetaData(props.resource.bucketId, props.resource.id, meta).then(() => {
-              props.setVideo({...props.resource, userMeta: meta });
-              setShowInfoModal(false)
-            })
-          } } 
-        />
-      </Modal>
-      
-      <Modal visible = { showDeleteDialog } onHide = { cancelDelete }>
-        <Dialog title = "Are you sure?">
-          <p>Do you want to delete: '{props.resource.userMeta.title}'</p>
-          <p>
-            <button className = "button-primary" onClick = { confirmDelete }>Yes</button>
-            <button onClick = { cancelDelete }>No / Cancel</button>
-          </p>
-        </Dialog>
-      </Modal>
-
-      <div className = "preview-menu">
-
-        <DropDown 
-          align = 'right' 
-          contentClassName = "dropdown-menu" 
-          toggleIcon = { <ImgWithAlt className = "preview-menu-icon" src="/icons/more.svg" /> } 
-          hideOnClick = {true} >
-          <MenuItem onClick = { () => { setShowInfoModal(true); props.onDialogOpen() } }>
-            <ImgWithAlt className="menu-icon" src="/icons/info.svg" />Info
-          </MenuItem>
-          <MenuItem href={`/editor/${props.resource.id}`}>
-            <ImgWithAlt className="menu-icon" src="/icons/edit.svg" />Fragments
-          </MenuItem>
-          <MenuItem onClick = { () => { setShowDeleteDialog(true); props.onDialogOpen() } }>
-            <ImgWithAlt className="menu-icon" src="/icons/delete.svg" />Delete
-          </MenuItem>
-        </DropDown>
-      </div>
-    </>
-  );
 }
 
 export default Preview
