@@ -1,17 +1,17 @@
 import { useState, useEffect } from "react";
 import { Api } from "../api/Api";
-import { Fragment, SearchResult } from "../api/Model";
+import { Clip, SearchResult } from "../api/Model";
 import FragmentsPlayer from "../components/common/FragmentsPlayer";
 
 const Compilation = () => {
 
-    const [fragments, setFragments] = useState<Array<Fragment>>([])
+    const [fragments, setFragments] = useState<Array<Clip>>([])
   
     useEffect(() => {
   
         Api.getMedias("", 24, 0).then(response => {
             const f = response as SearchResult
-            const frags = f.results.flatMap((v) => { return v.highlights })
+            const frags = f.results.flatMap((v) => { return v.clips })
             setFragments(frags)
         });
       }, []
